requirements.txt
# DarValue.ai Requirements
# Core dependencies for AI-powered real estate valuation platform

# Web Scraping
scrapy==2.11.0
beautifulsoup4==4.12.2
selenium==4.15.0
requests==2.31.0
lxml==4.9.3
html5lib==1.1

# Database
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
alembic==1.12.1

# Data Processing
pandas==2.1.3
numpy==1.25.2
polars==0.19.13

# Geospatial
geopandas==0.14.1
shapely==2.0.2
osmnx==1.6.0
folium==0.15.0
geopy==2.4.0
pyproj==3.6.1

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.1
lightgbm==4.1.0
catboost==1.2.2

# Computer Vision
opencv-python==4.8.1.78
pillow==10.1.0
tensorflow==2.14.0
torch==2.1.1
torchvision==0.16.1

# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.2

# Cloud Storage
boto3==1.34.0
google-cloud-storage==2.10.0

# Configuration & Environment
python-decouple==3.8
pyyaml==6.0.1
python-dotenv==1.0.0

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Utilities
tqdm==4.66.1
schedule==1.2.0
celery==5.3.4
redis==5.0.1